Added PreviewActivity to show preview
