+++ app/src/main/java/com/shekhargulati/missingkidtracker/MainActivity.java
@@ -2,21 +2,15 @@ package com.shekhargulati.missingkidtracker;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Matrix;
-import android.media.ExifInterface;
 import android.net.Uri;
+import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.ImageView;
 
 import java.io.File;
 import java.io.IOException;
@@ -103,7 +97,9 @@ public class MainActivity extends AppCompatActivity {
         if (resultCode == RESULT_OK) {
             Log.d(tag, String.format("Photo is successfully saved to [%s]", capturedPhotoPath));
             addPicToGallery();
-            setPic();
+            Intent previewIntent = new Intent(this, PreviewActivity.class);
+            previewIntent.putExtra(PreviewActivity.CAPTURED_PHOTO_PATH, capturedPhotoPath);
+            startActivity(previewIntent);
         }
     }
 
@@ -115,45 +111,6 @@ public class MainActivity extends AppCompatActivity {
         this.sendBroadcast(mediaScanIntent);
     }
 
-    private void setPic() {
-        final String tag = getString(R.string.app_name);
-        try {
-            ImageView imageView = (ImageView) this.findViewById(R.id.image_preview);
-            imageView.setVisibility(View.VISIBLE);
-            Bitmap bitmap = createScaledBitmap(capturedPhotoPath);
-            imageView.setImageBitmap(bitmap);
-        } catch (Exception e) {
-            Log.e(tag, "Error encountered while doing image preview", e);
-        }
-    }
 
-    public Bitmap createScaledBitmap(String pathName) throws IOException {
-        final BitmapFactory.Options opt = new BitmapFactory.Options();
-        opt.inSampleSize = 2;
-        opt.inJustDecodeBounds = false;
-        Bitmap bitmap = BitmapFactory.decodeFile(pathName, opt);
-        File file = new File(capturedPhotoPath);
-        Bitmap rotatedBitmap;
-        ExifInterface exif = new ExifInterface(file.getPath());
-        int rotation = exif.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL);
-        int rotationInDegrees = exifToDegrees(rotation);
-        Matrix matrix = new Matrix();
-        if (rotation != 0f) {
-            matrix.preRotate(rotationInDegrees);
-        }
-        rotatedBitmap = Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, true);
-        return rotatedBitmap;
-    }
-
-    private static int exifToDegrees(int exifOrientation) {
-        if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_90) {
-            return 90;
-        } else if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_180) {
-            return 180;
-        } else if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_270) {
-            return 270;
-        }
-        return 0;
-    }
 
 }
