+++ app/src/main/java/com/shekhargulati/missingkidtracker/PreviewActivity.java
@@ -1,0 +1,71 @@
+package com.shekhargulati.missingkidtracker;
+
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Matrix;
+import android.media.ExifInterface;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.ImageView;
+
+import java.io.File;
+import java.io.IOException;
+
+public class PreviewActivity extends AppCompatActivity {
+
+    public static final String CAPTURED_PHOTO_PATH = "capturedPhotoPath";
+    private String capturedPhotoPath;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_preview);
+        Intent intent = getIntent();
+        this.capturedPhotoPath = intent.getStringExtra(CAPTURED_PHOTO_PATH);
+        setPic();
+    }
+
+    private void setPic() {
+        final String tag = getString(R.string.app_name);
+        try {
+            ImageView imageView = (ImageView) this.findViewById(R.id.image_preview);
+            imageView.setVisibility(View.VISIBLE);
+            Bitmap bitmap = createScaledBitmap(capturedPhotoPath);
+            imageView.setImageBitmap(bitmap);
+        } catch (Exception e) {
+            Log.e(tag, "Error encountered while doing image preview", e);
+        }
+    }
+
+    public Bitmap createScaledBitmap(String pathName) throws IOException {
+        final BitmapFactory.Options opt = new BitmapFactory.Options();
+        opt.inSampleSize = 2;
+        opt.inJustDecodeBounds = false;
+        Bitmap bitmap = BitmapFactory.decodeFile(pathName, opt);
+        File file = new File(capturedPhotoPath);
+        Bitmap rotatedBitmap;
+        ExifInterface exif = new ExifInterface(file.getPath());
+        int rotation = exif.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL);
+        int rotationInDegrees = exifToDegrees(rotation);
+        Matrix matrix = new Matrix();
+        if (rotation != 0f) {
+            matrix.preRotate(rotationInDegrees);
+        }
+        rotatedBitmap = Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, true);
+        return rotatedBitmap;
+    }
+
+    private static int exifToDegrees(int exifOrientation) {
+        if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_90) {
+            return 90;
+        } else if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_180) {
+            return 180;
+        } else if (exifOrientation == ExifInterface.ORIENTATION_ROTATE_270) {
+            return 270;
+        }
+        return 0;
+    }
+}
